Use some helper macros to allow us to use PA4 and PC13 and so on where we imply the PORT A, B, C etc

+++ libraries/nrf24_hal_support/nrf24_hal_support.types.h
@@ -6,12 +6,13 @@ struct nrf24_hal_support_interface
 	 * address, the interrupt, CE and CS pins are also configured and must be in the same GPIO port as
 	 * the chosen SPI instance.
 	 * @param spi_base The base address of the SPI instance to be configured.
+	 * @param tim_base The base address of the timer instance to be used for CE pulse generation.
 	 * @param int_pin The GPIO pin number for the interrupt line (negative value means no interrupt).
 	 * @param ce_pin The GPIO pin number for the CE (Chip Enable) line.
 	 * @param cs_pin The GPIO pin number for the CSN (Chip Select Not) line.
 	 * @param device_ptr The returned configured device instance. 
 	*/
-	void(*Configure)(SPI_TypeDef * spi_base, TIM_TypeDef * tim_base, int32_t int_pin, uint32_t ext_int_id, uint32_t ce_pin, uint32_t cs_pin, NrfSpiDevice_ptr device_ptr, nrf_fault_handler handler);
+	void(*Configure)(SPI_TypeDef * spi_base, TIM_TypeDef * tim_base, int64_t int_pin, uint32_t ext_int_id, uint64_t ce_pin, uint64_t cs_pin, NrfSpiDevice_ptr device_ptr, nrf_fault_handler handler);
 	/** 
 	 Sets the NRF device's CE pin to low
 	 */
